Change registration validation to account validation
