+++ src/views/_layouts/connected.phtml
@@ -54,8 +54,8 @@
 
             <?php if (!$current_user->validated_at && $banner): ?>
                 <div class="layout__banner">
-                    <a class="banner__container banner__container--anchor" href="<?= url('registration validation') ?>">
-                        <?= _('You must validate your registration&nbsp;→') ?>
+                    <a class="banner__container banner__container--anchor" href="<?= url('account validation') ?>">
+                        <?= _('You must validate your account&nbsp;→') ?>
                     </a>
                 </div>
             <?php endif; ?>
