+++ src/views/mailers/users/account_validation_email.phtml
@@ -4,8 +4,8 @@
 
 <p>
     <?= _f('Thank you for your inscription on %s. To finish, you’ll have to confirm your email address. You just have to click on the following link, and you should be all good then:', url_full('home')) ?></br />
-    <a href="<?= url_full('registration validation', ['t' => $token]) ?>">
-        <?= url_full('registration validation', ['t' => $token]) ?>
+    <a href="<?= url_full('account validation', ['t' => $token]) ?>">
+        <?= url_full('account validation', ['t' => $token]) ?>
     </a>
 </p>
