+++ src/views/mailers/users/account_validation_email.txt
@@ -4,7 +4,7 @@
 <?= _f('Thank you for your inscription on %s. To finish, you’ll have to confirm your email address. You just have to click on the following link, and you should be all good then:', url_full('home')) ?>
 
 
-<?= url_full('registration validation', ['t' => $token]) ?>
+<?= url_full('account validation', ['t' => $token]) ?>
 
 
 <?= _('Have a nice day!') ?>
