+++ src/views/my/account/show.phtml
@@ -32,7 +32,7 @@
 
     <?php if (!$current_user->validated_at): ?>
         <p class="paragraph--centered">
-            <a class="anchor--action" href="<?= url('registration validation') ?>">
+            <a class="anchor--action" href="<?= url('account validation') ?>">
                 <?= _('Validate your account') ?>
             </a>
         </p>
